+++ src/dk/aau/cs/giraf/train/PictogramReceiver.java
@@ -1,5 +1,5 @@
 package dk.aau.cs.giraf.train;
 
 public interface PictogramReceiver {
-    public void receivePictograms(long[] pictogramIds, int requestCode);
+    public void receivePictograms(int[] pictogramIds, int requestCode);
 }
